requirements.txt
# Flexible pins for Python 3.11–3.13; run: pip install -r requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-groq>=0.1.3
langchain-community>=0.2.0
langgraph>=0.2.0
faiss-cpu>=1.9.0,<2
sentence-transformers>=2.7.0
rank-bm25==0.2.2
streamlit>=1.35.0
python-dotenv>=1.0.1
pypdf>=4.2.0
groq>=0.9.0
